pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "soc-triage-env"
version = "1.0.0"
description = "SOC Triage Environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "openenv-core>=0.2.0"
]

[project.scripts]
server = "server.app:start"
